requirements.txt
aiomysql==0.2.0
bson==0.5.10
fastapi==0.116.1
google-genai==1.27.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
kaleido==1.0.0
mysql-connector-python==9.4.0
pandas==2.3.1
pillow==11.3.0
plotly==6.2.0
pydantic==2.11.7
python-dotenv==1.1.1
python-telegram-bot==22.3
redis==6.2.0
requests==2.32.4
SQLAlchemy==2.0.41

